requirements.txt
torch>=1.7.1
torchvision>=0.8.2
transformers>=4.5.0
scikit-learn>=0.24.1
matplotlib>=3.3.4
numpy>=1.19.5
Pillow>=8.1.0
tqdm>=4.56.0
tensorboard>=2.4.1
nltk>=3.6.0
wordcloud>=1.8.1
pandas>=1.2.3 